requirements.txt
aiofiles==0.4.0
certifi==2019.6.16
chardet==3.0.4
Django==2.0
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.10.3
idna==2.8
jieba==0.39
mecache==1.3.0
mysqlclient==1.4.4
Pillow==6.1.0
pysolr==3.8.1
python-memcached==1.59
pytz==2019.2
qiniu==7.2.6
redis==3.3.8
requests==2.22.0
shortuuid==0.5.0
six==1.12.0
urllib3==1.25.3
Whoosh==2.7.4
